Migrate image_picker and firebase_storage to flutter 2 null_safety
