--- lib/resources/storage/storage_provider.dart
@@ -1,4 +1,0 @@
-/// Call the correct method depending on the platform used
-export 'unsupported_storage.dart'
-    if (dart.library.html) 'web_storage.dart'
-    if (dart.library.io) 'mobile_storage.dart';
